* develop(1105-sftp-support): adding SFTP support

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): fix SSM parameter import/export

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add SFTP server

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add Cognito identity provider details

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): add Cognito identity provider details

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): add Cognito identity provider details

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* Merge remote-tracking branch 'refs/remotes/origin/main' into 1105-sft…

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* Merge branch 'main' into 1105-sftp-support

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): fix SQS message attributes

* develop(1105-sftp-support): fix SQS message attributes

* develop(1105-sftp-support): fix SQS message attributes

* develop(1105-sftp-support): fix SQS message attributes

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): add FTP users protocols

* develop(1105-sftp-support): fix queue/topic import

* develop(1105-sftp-support): fix queue/topic import

* develop(1105-sftp-support): fix queue/topic import

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): fix queue/topic import

* develop(1105-sftp-support): fix queue/topic import

* develop(1105-sftp-support): fix queue/topic import

* develop(1105-sftp-support): fix awslocal

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

* Merge remote-tracking branch 'origin/1105-sftp-support' into 1105-sft…

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

+++ src/dto/include/awsmock/dto/sqs/model/QueueCounter.h
@@ -54,12 +54,12 @@ namespace AwsMock::Dto::SQS {
         /**
          * Visibility timeout
          */
-        int visibilityTimeout{};
+        long visibilityTimeout{};
 
         /**
          * Message delay
          */
-        int delay{};
+        long delay{};
 
         /**
          * Retention period
@@ -72,7 +72,7 @@ namespace AwsMock::Dto::SQS {
         long maxMessageSize{};
 
         /**
-         * Total size of message in the queue
+         * Total size of all messages in the queue
          */
         long size{};
